wip: [skip ci] refactoring config change listeners
