+++ cmd/manager/http.go
@@ -27,12 +27,13 @@ package main
 import (
 	"github.com/flomesh-io/fsm/pkg/commons"
 	"github.com/flomesh-io/fsm/pkg/config"
+	"github.com/flomesh-io/fsm/pkg/config/utils"
 	"github.com/flomesh-io/fsm/pkg/repo"
 	"os"
 )
 
 func setupHTTP(repoClient *repo.PipyRepoClient, mc *config.MeshConfig) {
-	if err := config.UpdateIngressHTTPConfig(commons.DefaultIngressBasePath, repoClient, mc); err != nil {
+	if err := utils.UpdateIngressHTTPConfig(commons.DefaultIngressBasePath, repoClient, mc); err != nil {
 		os.Exit(1)
 	}
 }
